adwords-tool/requirements.txt
Flask==3.1.1
google-ads==21.0.0
python-dotenv==1.0.0
Werkzeug==3.1.1

